requirements.txt
celery==3.1.17
Flask==0.10.1
Django==1.9
django-celery==3.1.17
djangorestframework==3.5.3
redis==2.10.5
